修改规则引擎，添加评分卡模式，规则树模式修改为必须一竿子捅到底才算整体命中,否则返回未命中

+++ src/main/java/com/efy/function/dto/order/CancelAllDto.java
@@ -19,6 +19,6 @@ public class CancelAllDto {
     private Integer failedCount;
     @JSONField(name = "next-id")
     @ApiModelProperty("下一个可以撤销的订单号，返回-1表示没有可以撤销的订单")
-    private Integer nextId;
+    private Long nextId;
 
 }
